Moved test_std outside test_common
