+++ tests/frame/std_test.py
@@ -1,0 +1,25 @@
+from typing import Any
+
+import narwhals.stable.v1 as nw
+from tests.utils import compare_dicts
+
+data = {"a": [1, 3, 2], "b": [4, 4, 6], "z": [7.0, 8, 9]}
+
+
+def test_std(constructor_eager: Any) -> None:
+    df = nw.from_native(constructor_eager(data))
+    result = df.select(
+        nw.col("a").std().alias("a_ddof_default"),
+        nw.col("a").std(ddof=1).alias("a_ddof_1"),
+        nw.col("a").std(ddof=0).alias("a_ddof_0"),
+        nw.col("b").std(ddof=2).alias("b_ddof_2"),
+        nw.col("z").std(ddof=0).alias("z_ddof_0"),
+    )
+    expected = {
+        "a_ddof_default": [1.0],
+        "a_ddof_1": [1.0],
+        "a_ddof_0": [0.816497],
+        "b_ddof_2": [1.632993],
+        "z_ddof_0": [0.816497],
+    }
+    compare_dicts(result, expected)
